add shared gulpfile
