+++ elements/rh-health-index/package.json
@@ -1,5 +1,9 @@
 {
   "name": "@rhelements/rh-health-index",
+  "rhelement": {
+    "className": "RhHealthIndex",
+    "elementName": "rh-health-index"
+  },
   "version": "1.0.0-prerelease.3",
   "publishConfig": {
     "access": "public"
@@ -11,7 +15,7 @@
   },
   "main": "index.js",
   "scripts": {
-    "build": "../../node_modules/.bin/gulp",
+    "build": "../../node_modules/.bin/gulp && ../../node_modules/.bin/prettier --ignore-path ../../.prettierignore --write '**/*.{js,json}'",
     "dev": "../../node_modules/.bin/gulp dev",
     "test": "../../node_modules/.bin/wct --configFile ../../wct.conf.json node_modules/@rhelements/rh-health-index/test/"
   },
