+++ elements/rh-health-index/rh-health-index.umd.js
@@ -1,1 +1,2 @@
-!function(e,n){if("function"==typeof define&&define.amd)define(["exports","../rhelement/rhelement.umd.js"],n);else if("undefined"!=typeof exports)n(exports,require("../rhelement/rhelement.umd.js"));else{var t={};n(t,e.rhelementUmd),e.rhHealthIndex=t}}(this,function(e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});var t,r=(t=n)&&t.__esModule?t:{default:t};var o=function(){function r(e,n){for(var t=0;t<n.length;t++){var r=n[t];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}return function(e,n,t){return n&&r(e.prototype,n),t&&r(e,t),e}}();var a=function(e){function i(){return function(e,n){if(!(e instanceof n))throw new TypeError("Cannot call a class as a function")}(this,i),function(e,n){if(!e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return!n||"object"!=typeof n&&"function"!=typeof n?e:n}(this,(i.__proto__||Object.getPrototypeOf(i)).call(this,i.tag))}return function(e,n){if("function"!=typeof n&&null!==n)throw new TypeError("Super expression must either be null or a function, not "+typeof n);e.prototype=Object.create(n&&n.prototype,{constructor:{value:e,enumerable:!1,writable:!0,configurable:!0}}),n&&(Object.setPrototypeOf?Object.setPrototypeOf(e,n):e.__proto__=n)}(i,r.default),o(i,[{key:"html",get:function(){return'\n<style>\n:host {\n  display: inline-flex;\n  align-items: center; }\n\n:host([hidden]) {\n  display: none; }\n\n.box-container {\n  display: flex;\n  border: 1px solid var(--rh-theme--color--surface--border, #dfdfdf);\n  margin-left: var(--rh-theme--container-spacer--xs, ); }\n  .box-container .box {\n    background: #fff;\n    width: 10px;\n    height: 20px;\n    border-right: 1px solid var(--rh-theme--color--surface--border, #dfdfdf); }\n    .box-container .box:last-child {\n      border-right: 0; }\n    .box-container .box.active.a {\n      background-color: #3f9c35; }\n    .box-container .box.active.b {\n      background-color: #92d400; }\n    .box-container .box.active.c {\n      background-color: #efaa00; }\n    .box-container .box.active.d {\n      background-color: #ec7a08; }\n    .box-container .box.active.e {\n      background-color: #cc0000; }\n    .box-container .box.active.f {\n      background-color: #a30000; }\n</style>\n<div id="healthIndex"></div>\n<div class="box-container">\n  <div class="box a"></div>\n  <div class="box b"></div>\n  <div class="box c"></div>\n  <div class="box d"></div>\n  <div class="box e"></div>\n  <div class="box f"></div>\n</div>'}},{key:"templateUrl",get:function(){return"rh-health-index.html"}},{key:"styleUrl",get:function(){return"rh-health-index.scss"}}],[{key:"tag",get:function(){return"rh-health-index"}},{key:"observedAttributes",get:function(){return["health-index"]}}]),o(i,[{key:"attributeChangedCallback",value:function(e,n,t){var r=t.toLowerCase(),o=t.toUpperCase(),a=[].concat(function(e){if(Array.isArray(e)){for(var n=0,t=Array(e.length);n<e.length;n++)t[n]=e[n];return t}return Array.from(e)}(this.shadowRoot.querySelectorAll(".box")));this.innerHTML=o,this.shadowRoot.querySelector("#healthIndex").innerText=o,a.forEach(function(e){e.classList.contains(r)?e.classList.add("active"):e.classList.remove("active")}),this.shadowRoot.querySelector(".box.active")||console.warn(i.tag+": a valid health-index was not provided. Please use A, B, C, D, E, or F")}}]),i}();r.default.create(a),e.default=a});
+!function(e,n){"object"==typeof exports&&"undefined"!=typeof module?module.exports=n(require("../rhelement/rhelement.umd.js")):"function"==typeof define&&define.amd?define(["../rhelement/rhelement.umd.js"],n):e.RhHealthIndex=n(e.RHElement)}(this,function(n){"use strict";n=n&&n.hasOwnProperty("default")?n.default:n;var t=function(){function o(e,n){for(var t=0;t<n.length;t++){var o=n[t];o.enumerable=o.enumerable||!1,o.configurable=!0,"value"in o&&(o.writable=!0),Object.defineProperty(e,o.key,o)}}return function(e,n,t){return n&&o(e.prototype,n),t&&o(e,t),e}}(),e=function(e){function a(){return function(e,n){if(!(e instanceof n))throw new TypeError("Cannot call a class as a function")}(this,a),function(e,n){if(!e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return!n||"object"!=typeof n&&"function"!=typeof n?e:n}(this,(a.__proto__||Object.getPrototypeOf(a)).call(this,a.tag))}return function(e,n){if("function"!=typeof n&&null!==n)throw new TypeError("Super expression must either be null or a function, not "+typeof n);e.prototype=Object.create(n&&n.prototype,{constructor:{value:e,enumerable:!1,writable:!0,configurable:!0}}),n&&(Object.setPrototypeOf?Object.setPrototypeOf(e,n):e.__proto__=n)}(a,n),t(a,[{key:"html",get:function(){return'\n<style>\n:host {\n  display: inline-flex;\n  align-items: center; }\n\n:host([hidden]) {\n  display: none; }\n\n.box-container {\n  display: flex;\n  border: 1px solid var(--rh-theme--color--surface--border, #dfdfdf);\n  margin-left: var(--rh-theme--container-spacer--xs, ); }\n  .box-container .box {\n    background: #fff;\n    width: 10px;\n    height: 20px;\n    border-right: 1px solid var(--rh-theme--color--surface--border, #dfdfdf); }\n    .box-container .box:last-child {\n      border-right: 0; }\n    .box-container .box.active.a {\n      background-color: #3f9c35; }\n    .box-container .box.active.b {\n      background-color: #92d400; }\n    .box-container .box.active.c {\n      background-color: #efaa00; }\n    .box-container .box.active.d {\n      background-color: #ec7a08; }\n    .box-container .box.active.e {\n      background-color: #cc0000; }\n    .box-container .box.active.f {\n      background-color: #a30000; }\n</style>\n<div id="healthIndex"></div>\n<div class="box-container">\n  <div class="box a"></div>\n  <div class="box b"></div>\n  <div class="box c"></div>\n  <div class="box d"></div>\n  <div class="box e"></div>\n  <div class="box f"></div>\n</div>'}},{key:"templateUrl",get:function(){return"rh-health-index.html"}},{key:"styleUrl",get:function(){return"rh-health-index.scss"}}],[{key:"tag",get:function(){return"rh-health-index"}},{key:"observedAttributes",get:function(){return["health-index"]}}]),t(a,[{key:"attributeChangedCallback",value:function(e,n,t){var o=t.toLowerCase(),r=t.toUpperCase(),i=[].concat(function(e){if(Array.isArray(e)){for(var n=0,t=Array(e.length);n<e.length;n++)t[n]=e[n];return t}return Array.from(e)}(this.shadowRoot.querySelectorAll(".box")));this.innerHTML=r,this.shadowRoot.querySelector("#healthIndex").innerText=r,i.forEach(function(e){e.classList.contains(o)?e.classList.add("active"):e.classList.remove("active")}),this.shadowRoot.querySelector(".box.active")||console.warn(a.tag+": a valid health-index was not provided. Please use A, B, C, D, E, or F")}}]),a}();return n.create(e),e});
+//# sourceMappingURL=rh-health-index.umd.js.map
