+++ elements/rh-health-index/rollup.config.js
@@ -1,0 +1,5 @@
+// rollup.config.js
+import configFactory from "../../scripts/rollup.config.factory.js";
+import rhelementPackage from "./package.json";
+
+export default configFactory(rhelementPackage.rhelement);
