+++ elements/rh-icon/gulpfile.js
@@ -1,95 +1,14 @@
-const path = require("path");
-const fs = require("fs");
-const del = require("del");
+// rollup.config.js
+const gulpFactory = require("../../scripts/gulpfile.factory.js");
+const rhelementPackage = require("./package.json");
 
+const fs = require("fs");
 const gulp = require("gulp");
 const rename = require("gulp-rename");
 const replace = require("gulp-replace");
-const stripCssComments = require("strip-css-comments");
-const trim = require("gulp-trim");
-const decomment = require("decomment");
-const sass = require("node-sass");
-const shell = require("gulp-shell");
-
-// Custom for the rh-icon component
 const svgSprite = require("gulp-svg-sprite");
 
-gulp.task("compile", () => {
-  return gulp
-    .src("./rh-button.js")
-    .pipe(
-      replace(
-        /^(import .*?)(['"]\.\.\/(?!\.\.\/).*)(\.js['"];)$/gm,
-        "$1$2.umd$3"
-      )
-    )
-    .pipe(
-      rename({
-        suffix: ".umd"
-      })
-    )
-    .pipe(gulp.dest("./"));
-});
-
-gulp.task("merge", () => {
-  return gulp
-    .src("./src/rh-icon.js")
-    .pipe(
-      replace(
-        /extends\s+RHElement\s+{/g,
-        (classStatement, character, jsFile) => {
-          // extract the templateUrl and styleUrl with regex.  Would prefer to do
-          // this by require'ing rh-something.js and asking it directly, but without
-          // node.js support for ES modules, we're stuck with this.
-          const oneLineFile = jsFile
-            .slice(character)
-            .split("\n")
-            .join(" ");
-          const [
-            ,
-            templateUrl
-          ] = /get\s+templateUrl\([^)]*\)\s*{\s*return\s+"([^"]+)"/.exec(
-            oneLineFile
-          );
-
-          let html = fs
-            .readFileSync(path.join("./src", templateUrl))
-            .toString()
-            .trim();
-
-          html = decomment(html);
-
-          const [
-            ,
-            styleUrl
-          ] = /get\s+styleUrl\([^)]*\)\s*{\s*return\s+"([^"]+)"/.exec(
-            oneLineFile
-          );
-
-          const styleFilePath = path.join("./src", styleUrl);
-
-          let cssResult = sass.renderSync({
-            file: styleFilePath
-          }).css;
-
-          cssResult = stripCssComments(cssResult).trim();
-
-          return `${classStatement}
-  get html() {
-    return \`
-<style>
-${cssResult}
-</style>
-${html}\`;
-  }
-`;
-        }
-      )
-    )
-    .pipe(gulp.dest("./"));
-});
-
-// Custom for the rh-icon component
+// Custom gulp tasks for rh-icon
 gulp.task("svgSprite", function() {
   return gulp
     .src("./svg/*.svg")
@@ -123,33 +42,8 @@ gulp.task("stuffSprite", () => {
     .pipe(gulp.dest("./"));
 });
 
-gulp.task("compile", () => {
-  return gulp
-    .src(["./rh-icon.js"])
-    .pipe(
-      replace(
-        /^(import .*?)(['"]\.\.\/(?!\.\.\/).*)(\.js['"];)$/gm,
-        "$1$2.umd$3"
-      )
-    )
-    .pipe(
-      rename({
-        suffix: ".umd"
-      })
-    )
-    .pipe(gulp.dest("./"));
+// call the central gulp build, and pass in the custom tasks to be run pre-bundle
+gulpFactory({
+  ...rhelementPackage.rhelement,
+  taskHooks: { prebundle: ["svgSprite", "stuffSprite"] }
 });
-
-gulp.task("watch", () => {
-  return gulp.watch("./src/*", gulp.series("build"));
-});
-
-gulp.task("svgs", gulp.series("svgSprite", "stuffSprite"));
-
-gulp.task("bundle", shell.task("../../node_modules/.bin/rollup -c"));
-
-gulp.task("build", gulp.series("merge", "svgs", "compile", "bundle"));
-
-gulp.task("default", gulp.series("build"));
-
-gulp.task("dev", gulp.series("build", "watch"));
